Documentation for Monotype's OpenType Layout Source File Format
Beginning in 1999 Monotype needed the ability to create, view and debug
fonts with OpenType Layout tables. A text representation of the OpenType
tables (GSUB, GPOS, and GDEF) was created that follows the actual
structure of the tables so that these source files could be compiled
into binary format and vice versa. This document describes the format of
such source files. With appropriate tools such files can be converted to
binary data and inserted into fonts.